+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - Added option to Controller to permit caching unsafe HTTP methods.
+- Support AWS S3 storages. (#164)
+- Move `typing_extensions` from requirements.txt to pyproject.toml. (#161)
 
 ## 0.0.21 (29th December, 2023)