+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.0.22 (TODO)
 
 - Make `FileStorage` to check staleness of all cache files with set interval. (#169)
 - Support AWS S3 storages. (#164)